+++ dejafu/Test/DejaFu/Conc/Internal/Program.hs
@@ -92,11 +92,9 @@ instance (pty ~ Basic, Monad n) => C.MonadConc (Program pty n) where
 
   forkWithUnmaskN   n ma = ModelConc (AFork n (\umask -> runModelConc (ma umask) (\_ -> AStop (pure ()))))
   forkOnWithUnmaskN n _  = C.forkWithUnmaskN n
-  forkOSWithUnmaskN n ma
-    | C.rtsSupportsBoundThreads =
-      ModelConc (AForkOS n (\umask -> runModelConc (ma umask) (\_ -> AStop (pure ()))))
-    | otherwise = fail "RTS doesn't support multiple OS threads (use ghc -threaded when linking)"
+  forkOSWithUnmaskN n ma = ModelConc (AForkOS n (\umask -> runModelConc (ma umask) (\_ -> AStop (pure ()))))
 
+  supportsBoundThreads = ModelConc ASupportsBoundThreads
   isCurrentThreadBound = ModelConc AIsBound
 
   -- This implementation lies and returns 2 until a value is set. This
@@ -172,8 +170,8 @@ instance (pty ~ Basic, Monad n) => C.MonadConc (Program pty n) where
 -- nonexistent thread. In either of those cases, the computation will
 -- be halted.
 --
--- @since 2.0.0.0
-runConcurrent :: C.MonadConc n
+-- @since 2.1.0.0
+runConcurrent :: MonadDejaFu n
   => Scheduler s
   -> MemType
   -> s
@@ -181,7 +179,7 @@ runConcurrent :: C.MonadConc n
   -> n (Either Condition a, s, Trace)
 runConcurrent sched memtype s ma@(ModelConc _) = do
   res <- runConcurrency [] False sched memtype s initialIdSource 2 ma
-  out <- efromJust <$> C.readIORef (finalRef res)
+  out <- efromJust <$> readRef (finalRef res)
   pure ( out
        , cSchedState (finalContext res)
        , F.toList (finalTrace res)
@@ -252,9 +250,9 @@ runConcurrent sched memtype s ma = recordSnapshot ma >>= \case
 --   (liftIO . readIORef)
 -- @
 --
--- @since 2.0.0.0
+-- @since 2.1.0.0
 recordSnapshot
-  :: C.MonadConc n
+  :: MonadDejaFu n
   => Program pty n a
   -> n (Maybe (Either Condition (Snapshot pty n a), Trace))
 recordSnapshot ModelConc{..} = pure Nothing
@@ -267,9 +265,9 @@ recordSnapshot WithSetupAndTeardown{..} =
 
 -- | Runs a program with snapshotted setup to completion.
 --
--- @since 2.0.0.0
+-- @since 2.1.0.0
 runSnapshot
-  :: C.MonadConc n
+  :: MonadDejaFu n
   => Scheduler s
   -> MemType
   -> s
@@ -278,7 +276,7 @@ runSnapshot
 runSnapshot sched memtype s (WS SimpleSnapshot{..}) = do
   let context = fromSnapContext s snapContext
   CResult{..} <- runConcurrencyWithSnapshot sched memtype context snapRestore snapNext
-  out <- efromJust <$> C.readIORef finalRef
+  out <- efromJust <$> readRef finalRef
   pure ( out
        , cSchedState finalContext
        , F.toList finalTrace
@@ -287,9 +285,9 @@ runSnapshot sched memtype s (WSAT SimpleSnapshot{..} teardown) = do
   let context = fromSnapContext s snapContext
   intermediateResult <- runConcurrencyWithSnapshot sched memtype context snapRestore snapNext
   let idsrc = cIdSource (finalContext intermediateResult)
-  out1 <- efromJust <$> C.readIORef (finalRef intermediateResult)
+  out1 <- efromJust <$> readRef (finalRef intermediateResult)
   teardownResult <- simpleRunConcurrency False idsrc (teardown out1)
-  out2 <- efromJust <$> C.readIORef (finalRef teardownResult)
+  out2 <- efromJust <$> readRef (finalRef teardownResult)
   pure ( out2
        , cSchedState (finalContext intermediateResult)
        , F.toList (finalTrace intermediateResult)
@@ -328,7 +326,7 @@ threadsFromSnapshot snap = (initialThread : runnable, blocked) where
 -- | 'recordSnapshot' implemented generically.
 --
 -- Throws an error if the snapshot could not be produced.
-defaultRecordSnapshot :: C.MonadConc n
+defaultRecordSnapshot :: MonadDejaFu n
   => (SimpleSnapshot n a -> x -> snap)
   -> ModelConc n x
   -> (x -> ModelConc n a)
@@ -336,7 +334,7 @@ defaultRecordSnapshot :: C.MonadConc n
 defaultRecordSnapshot mkSnapshot setup program = do
   CResult{..} <- simpleRunConcurrency True initialIdSource setup
   let trc = F.toList finalTrace
-  out <- C.readIORef finalRef
+  out <- readRef finalRef
   pure . Just $ case out of
     Just (Right a) ->
       let snap = mkSnapshot (SimpleSnapshot finalContext finalRestore (program a)) a
@@ -351,7 +349,7 @@ defaultRecordSnapshot mkSnapshot setup program = do
 
 -- | Run a concurrent program with a deterministic scheduler in
 -- snapshotting or non-snapshotting mode.
-simpleRunConcurrency ::(C.MonadConc n, HasCallStack)
+simpleRunConcurrency ::(MonadDejaFu n, HasCallStack)
   => Bool
   -> IdSource
   -> ModelConc n a