+++ dejafu/Test/DejaFu/SCT.hs
@@ -1,5 +1,3 @@
-{-# OPTIONS_GHC -Wno-deprecations #-}
-
 -- |
 -- Module      : Test.DejaFu.SCT
 -- Copyright   : (c) 2015--2019 Michael Walker
@@ -26,7 +24,6 @@ module Test.DejaFu.SCT
 
 import           Control.Applicative               ((<|>))
 import           Control.DeepSeq                   (NFData(..), force)
-import           Control.Monad.Conc.Class          (MonadConc)
 import           Data.List                         (foldl')
 import qualified Data.Map.Strict                   as M
 import           Data.Maybe                        (fromMaybe)
@@ -51,8 +48,8 @@ import           Test.DejaFu.Utils
 -- The exact executions tried, and the order in which results are
 -- found, is unspecified and may change between releases.
 --
--- @since 2.0.0.0
-runSCT :: MonadConc n
+-- @since 2.1.0.0
+runSCT :: MonadDejaFu n
   => Way
   -- ^ How to run the concurrent program.
   -> MemType
@@ -64,8 +61,8 @@ runSCT way = runSCTWithSettings . fromWayAndMemType way
 
 -- | Return the set of results of a concurrent program.
 --
--- @since 2.0.0.0
-resultsSet :: (MonadConc n, Ord a)
+-- @since 2.1.0.0
+resultsSet :: (MonadDejaFu n, Ord a)
   => Way
   -- ^ How to run the concurrent program.
   -> MemType
@@ -83,8 +80,8 @@ resultsSet way = resultsSetWithSettings . fromWayAndMemType way
 -- The exact executions tried, and the order in which results are
 -- found, is unspecified and may change between releases.
 --
--- @since 2.0.0.0
-runSCT' :: (MonadConc n, NFData a)
+-- @since 2.1.0.0
+runSCT' :: (MonadDejaFu n, NFData a)
   => Way
   -- ^ How to run the concurrent program.
   -> MemType
@@ -99,8 +96,8 @@ runSCT' way = runSCTWithSettings' . fromWayAndMemType way
 -- Demanding the result of this will force it to normal form, which
 -- may be more efficient in some situations.
 --
--- @since 2.0.0.0
-resultsSet' :: (MonadConc n, Ord a, NFData a)
+-- @since 2.1.0.0
+resultsSet' :: (MonadDejaFu n, Ord a, NFData a)
   => Way
   -- ^ How to run the concurrent program.
   -> MemType
@@ -118,8 +115,8 @@ resultsSet' way = resultsSetWithSettings' . fromWayAndMemType way
 -- The exact executions tried, and the order in which results are
 -- found, is unspecified and may change between releases.
 --
--- @since 2.0.0.0
-runSCTWithSettings :: MonadConc n
+-- @since 2.1.0.0
+runSCTWithSettings :: MonadDejaFu n
   => Settings n a
   -- ^ The SCT settings.
   -> Program pty n a
@@ -159,8 +156,8 @@ runSCTWithSettings settings conc = case _way settings of
 
 -- | A variant of 'resultsSet' which takes a 'Settings' record.
 --
--- @since 2.0.0.0
-resultsSetWithSettings :: (MonadConc n, Ord a)
+-- @since 2.1.0.0
+resultsSetWithSettings :: (MonadDejaFu n, Ord a)
   => Settings n a
   -- ^ The SCT settings.
   -> Program pty n a
@@ -178,8 +175,8 @@ resultsSetWithSettings settings conc =
 -- The exact executions tried, and the order in which results are
 -- found, is unspecified and may change between releases.
 --
--- @since 2.0.0.0
-runSCTWithSettings' :: (MonadConc n, NFData a)
+-- @since 2.1.0.0
+runSCTWithSettings' :: (MonadDejaFu n, NFData a)
   => Settings n a
   -- ^ The SCT settings.
   -> Program pty n a
@@ -194,8 +191,8 @@ runSCTWithSettings' settings conc = do
 -- Demanding the result of this will force it to normal form, which
 -- may be more efficient in some situations.
 --
--- @since 2.0.0.0
-resultsSetWithSettings' :: (MonadConc n, Ord a, NFData a)
+-- @since 2.1.0.0
+resultsSetWithSettings' :: (MonadDejaFu n, Ord a, NFData a)
   => Settings n a
   -- ^ The SCT settings.
   -> Program pty n a